backend/requirements.txt
flask
flask-cors
python-dotenv
flask-pymongo
flask-jwt-extended
numpy
pandas
sentence-transformers
faiss-cpu
scikit-learn
deep-translator
requests
